+++ src/TextInputStyle.lua
@@ -1,0 +1,6 @@
+--!strict
+
+return {
+	Short = 1,
+	Paragraph = 2,
+}